Better routes for school confirmation
