+++ app/views/lead_providers/confirm_schools/show.html.erb
